Refactor: external API 서비스를 도메인별로 분리 - auth/ 폴더: Google OAuth 관련 API - system/ 폴더: 백엔드 시스템 API (헬스체크, 버전) - 각 도메인별 index.ts로 재export - import 경로를 도메인 기반으로 변경 - 확장성 향상 및 유지보수성 개선
