+++ src/services/external/auth/index.ts
@@ -1,0 +1,5 @@
+// Auth 도메인 외부 API 서비스들
+// Google OAuth, 사용자 관리 등 인증 관련 외부 API
+
+export * from './googleApi';
+// export * from './userApi'; // 향후 추가 예정
