+++ src/services/external/index.ts
@@ -1,0 +1,5 @@
+// External API 서비스들
+// 모든 외부 API 서비스를 중앙에서 관리
+
+export * from './auth';
+export * from './system';
