+++ src/services/external/system/index.ts
@@ -1,0 +1,4 @@
+// System 도메인 외부 API 서비스들
+// 백엔드 시스템 관련 API (헬스체크, 버전 등)
+
+export * from './systemApi';
